+++ laps.php
@@ -20,6 +20,10 @@ if ( ! class_exists( '\Rarst\Laps\Plugin' ) ) {
 	trigger_error( 'Laps not fully installed! Please install with Composer or download full release archive.', E_USER_ERROR );
 }
 
+if ( PHP_VERSION_ID < 70103 ) {
+	trigger_error( 'Laps requires PHP version >=7.1.3.', E_USER_ERROR );
+}
+
 $laps = new \Rarst\Laps\Plugin();
 
 $laps->run();