+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - handling of reoccurring hook events
 - tracking of load events when network–activated
 
+### Deprecated
+- dedicated classes for extension events.
+
 ## 3.1 - 2019-01-22
 
 ### Added